+++ mason_bricks/talker/hooks/pre_gen.dart
@@ -1,0 +1,14 @@
+import 'package:mason/mason.dart';
+import 'package:talker/talker.dart';
+
+void run(HookContext context) {
+  context.vars['lang'] = 'dart';
+  // var talkerSettings = TalkerSettings();
+
+  // final useDefaultSettings = context.vars['default_settings'];
+  // if (!useDefaultSettings) {
+  //   context.vars['use_history'] = talkerSettings.useHistory;
+  //   context.vars['use_console_logs'] = talkerSettings.useConsoleLogs;
+  //   context.vars['max_history_items'] = talkerSettings.maxHistoryItems;
+  // }
+}