+++ packages/talker_flutter/lib/src/ui/talker_monitor/talker_monitor_typed_logs_screen.dart
@@ -17,7 +17,7 @@ class TalkerMonitorTypedLogsScreen extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      backgroundColor: theme.backgroudColor,
+      backgroundColor: theme.backgroundColor,
       appBar: AppBar(
         title: Text('Talker Monitor $typeName'),
       ),