+++ packages/talker_flutter/lib/src/ui/talker_wrapper/talker_wrapper.dart
@@ -1,14 +1,13 @@
 import 'package:flutter/material.dart';
-import 'package:talker_flutter/src/ui/widgets/snackbar.dart';
 import 'package:talker_flutter/talker_flutter.dart';
 
 part 'talker_wrapper_options.dart';
 
 /// Widget to wrap an screen or the entire application
-/// in [Talker] data listerning
+/// in [Talker] data listening
 ///
 /// This is used to display error
-/// and exeption messages via [TalkerListener]
+/// and exception messages via [TalkerListener]
 ///
 /// [TalkerWrapperOptions] options used to customize and filtering
 /// error and exceptions messages output
@@ -64,7 +63,7 @@ class TalkerWrapper extends StatelessWidget {
   }
 
   static void showAlert(BuildContext context, Widget content) {
-    ScaffoldMessenger.of(context).showSnackBar(
+    ScaffoldMessenger.maybeOf(context)?.showSnackBar(
       SnackBar(
         margin: EdgeInsets.zero,
         behavior: SnackBarBehavior.floating,