+++ packages/talker/CHANGELOG.md
@@ -1,3 +1,32 @@
+# 3.1.1
+- Make 100% tests coverage
+- Add Contributors section in project README
+
+# 3.1.0
+- Rename (fix typo) **LoggerFormater** -> **LoggerFormatter** 
+- Rename (fix typo) Talker field **loggerFormater** -> **loggerFormatter** 
+- Fix internal and docs typos 
+- Update **talker_logger** version to 3.1.0
+
+Thanks to [wcoder](https://github.com/wcoder)
+
+# 3.0.5
+- Fix StackTrace printning when exits in **TalkerLog**
+
+Thanks to [IlyaZadyabin](https://github.com/IlyaZadyabin)
+
+# 3.0.4
+- - Add topics in pubspec.yaml
+
+# 3.0.3
+- Update **talker_logger** version to 3.0.4
+
+# 3.0.2
+- Update **talker_logger** version to 3.0.3
+
+# 3.0.1
+- Update sdk version to **'>=2.15.0 <4.0.0'**
+
 # 3.0.0
 
 **Lighter, simpler, more powerful**